third_party/blink/web_tests/svg/as-image/svg-image-usecounters-in-css-image.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
    #image {
        background-image: url('resources/animated-rect-color.svg');
        width: 100px;
        height: 100px;
    }
</style>
<div id="image"></div>
<script>
// From third_party/blink/public/mojom/use_counter/metrics/web_feature.mojom
var SVGSMILAnimationInImageRegardlessOfCache = 768;
var SVGImage = 5492;

window.jsTestIsAsync = true;

shouldBeFalse("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
shouldBeFalse("internals.isUseCounted(document, SVGImage)");

window.onload = function() {
    shouldBeTrue("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
    shouldBeTrue("internals.isUseCounted(document, SVGImage)");
    finishJSTest();
}
</script>
</html>
