third_party/blink/web_tests/svg/as-image/svg-image-usecounters-in-css-image-expected.txt

PASS internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is false
PASS internals.isUseCounted(document, SVGImage) is false
PASS internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is true
PASS internals.isUseCounted(document, SVGImage) is true
PASS successfullyParsed is true

TEST COMPLETE
